video-subtitle-extractor/requirements.txt
yt-dlp
webvtt-py
pysrt
